refactor(transitive-overlay): Extract route layer code to separate file.
